requirements.txt
# fastapi==0.104.1
graphql-core==3.2.3
strawberry-graphql==0.215.1
# uvicorn==0.24.0.post1
# uvloop==0.19.0
numpy==1.24.2
opencv-python-headless==4.8.1.78
onnx==1.15.0
onnxruntime==1.16.3
Jinja2==3.1.2

